+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This document records all notable changes to the OpenBBS codebase.
 This project uses the semantic versioning scheme.
 
 
+**Version 0.4.0**
+-----------------
+* Implemented countermeasures for control character injection.
+* Fixed thread post formatting.
+
+
 **Version 0.3.0**
 -----------------
 * Changed name of project to "OpenBBS."